requirements.txt
absl-py==2.2.2
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
blake3==1.0.5
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.33.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.0
ftfy==6.3.1
gguf==0.16.3
google==3.0.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
llguidance==0.7.24
llvmlite==0.44.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
natsort==8.4.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==1.24.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.82.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prometheus_client==0.22.0
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.52.3
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vllm==0.8.5.post1
wcwidth==0.2.13
Werkzeug==3.1.3
xgrammar==0.1.19
yarl==1.20.0
zmq==0.0.0
